requirements.txt
Flask
speechrecognition
deep-translator
transformers
torch
torchaudio
librosa
ffmpeg-python
